De-Conflate Prefixes
There are two kinds: the kind found in the scene lists and type given to
GDAL.  This distinction is important because `https://...` reads do not
work with the GDAL 2.2.3 that I just compiled (the following error is
produced):

```
ERROR 1: An error occurred while creating a virtual connection to the DAP server:
Error while reading the URL:
https://s3-us-west-2.amazonaws.com/landsat-pds/c1/L8/139/045/LC08_L1TP_139045_20170304_20170316_01_T1/LC08_L1TP_139045_2170304_20170316_01_T1_B4.TIF.ver.
The OPeNDAP server returned the following message:
Not Found: The data source or server could not be found.
Often this means that the OPeNDAP server is missing or needs attention;
Please contact the server administrator.
```